short-term-pipeline/requirements.txt
requests
pandas
python-dotenv
pymssql
pytest
lambda-multiprocessing